fix: slice 56d -- R1 ROOT-CAUSED + FIXED: a ONE-BYTE SHEAR in AF_UNIX >=64KB messages (silent 65535 clamp + full-length return) misframed Mojo/ipcz and made the watch-page renderer cleanly exit; media pipeline now UNLOCKED (10 segment fetches vs 2)
The chring instrument (per-thread-group ring of channel ops dumped at
renderer exit_group) captured the death prelude in one run: recvmsg 4096 +
61439 = 65535 bytes immediately before every clean exit. Mechanism: sendmsg
totals are honestly capped at 65536 (SYS_MAX_RW), but unix_enqueue_fds then
silently clamped 65536 -> 65535 while sock_unix_send_fds reported the full
count -- the receiver's byte stream lost exactly ONE BYTE per >=64KB
message, shearing ipcz framing; chrome's channel-error path then cleanly
shut the renderer down (and on non-fatal endpoints just silenced the
channel -- the probable true cause of the runs-3/4 message-flow stops).
The slice-53 "large IPC" exoneration had only tested up to 48KB -- below
the cliff.

Fix: sock_dgram.len + sock.tail_off u16 -> u32, whole-message enqueue, loud
8MiB sanity refusal instead of any silent cap. defboot clean. Run 10: the
clean-exit death is GONE, the renderer runs 2.3x further (62,935 syscalls),
issues 10 media requests with 10 responses (was 2/2), paints frames, and
the AF_UNIX dequeue checksum reports zero corruption. Remaining wall (next
arc, bounded): a user-mode wild-jump crash at ~30s -- the slice-50
0x208c13a corruption class reappearing further downstream; attack with the
[pgj]/[pfres]/g_tlbq toolset.

Also: chromewin --vmodule probe (release chrome has no VLOGs there; kept
harmless), probe cap 12->36 already in.

+++ tobyOS/include/tobyos/socket.h
@@ -87,7 +87,14 @@ struct file;
 struct sock_dgram {
     uint32_t  src_ip;          /* network byte order */
     uint16_t  src_port;        /* network byte order */
-    uint16_t  len;
+    /* Slice 56d: uint16_t until 2026-07-28 -- and unix_enqueue_fds SILENTLY
+     * TRUNCATED any AF_UNIX message > 65535 bytes while telling the sender it
+     * all went out. Chrome's browser->renderer Mojo channel messages cross
+     * 64KB right at player start; the sheared byte stream failed ipcz
+     * validation and the renderer cleanly exit(0)'d (the slice-56 R1 wall;
+     * measured as a 4096+61439=65535-byte read immediately before every
+     * death). 32 bits + no truncation. */
+    uint32_t  len;
     uint8_t  *payload;         /* kmalloc'd; freed on dequeue */
     struct file *fds[SOCK_SCM_MAX_FDS];
     uint8_t   nfds;
@@ -118,10 +125,11 @@ struct sock {
     uint8_t         tail;
     uint8_t         count;
     uint16_t        dropped;
-    uint16_t        tail_off;          /* AF_UNIX stream: bytes already consumed
+    uint32_t        tail_off;          /* AF_UNIX stream: bytes already consumed
                                         * from the head dgram (partial reads leave
                                         * the remainder queued, vs SEQPACKET which
-                                        * would drop it) -- X11 is a byte stream. */
+                                        * would drop it) -- X11 is a byte stream.
+                                        * u32: must span sock_dgram.len (slice 56d). */
     uint8_t         x_server;          /* 1 = in-kernel fake-X-server loopback:
                                         * writes are fed to the X handshake
                                         * responder, which enqueues replies back
